+++ servers/api/models/layout.js
@@ -19,7 +19,11 @@ module.exports = function (sequelize, DataTypes) {
             getAssociations: function () {
                 return [];
             },
-            sortOptions: {}
+            sortOptions: [
+                'createdAt',
+                'updatedAt',
+                'title'
+            ]
         }
     });
     return layout;