+++ servers/api/models/video.js
@@ -29,7 +29,13 @@ module.exports = function (sequelize, DataTypes) {
             getAssociations: function () {
                 return ['video'];
             },
-            sortOptions: {}
+            sortOptions: [
+                'createdAt',
+                'updatedAt',
+                'startTime',
+                'userId',
+                'host'
+            ]
         },
         hooks: {
             beforeCreate: function (video, options, callback) {