+++ servers/api/routes/models/comment.js
@@ -39,7 +39,6 @@ module.exports = {
     },
     resultModel: Joi.object(basicModel).options({className: resourceName}),
 
-
     scopes: {
         create: 'trusted',
         read: 'basic',